refactor: member 패키지 Response 객체들 패키지 수정
