+++ src/main/java/dandi/dandi/member/application/port/in/MemberUseCase.java
@@ -1,8 +1,5 @@
 package dandi.dandi.member.application.port.in;
 
-import dandi.dandi.member.web.dto.out.MemberInfoResponse;
-import dandi.dandi.member.web.dto.out.NicknameDuplicationCheckResponse;
-
 public interface MemberUseCase {
 
     MemberInfoResponse findMemberInfo(Long memberId);
