+++ src/main/java/dandi/dandi/member/application/service/ProfileImageService.java
@@ -3,11 +3,11 @@ package dandi.dandi.member.application.service;
 import com.amazonaws.SdkClientException;
 import dandi.dandi.auth.exception.UnauthorizedException;
 import dandi.dandi.image.exception.ImageUploadFailedException;
+import dandi.dandi.member.application.port.in.ProfileImageUpdateResponse;
 import dandi.dandi.member.application.port.in.ProfileImageUseCase;
 import dandi.dandi.member.application.port.out.MemberPersistencePort;
 import dandi.dandi.member.application.port.out.ProfileImageUploader;
 import dandi.dandi.member.domain.Member;
-import dandi.dandi.member.web.dto.out.ProfileImageUpdateResponse;
 import java.io.IOException;
 import java.util.UUID;
 import org.slf4j.Logger;
